requirements.txt
Flask
flask-cors
fpdf
python-dotenv
Pillow